Moved all the utils related to getFailure tool moved to failure-utils
